refactor(cli): group watch action files under cli/actions/watch
Move watchAction.ts and watchIgnore.ts (and their test) into a dedicated
cli/actions/watch/ directory so the multi-file watch feature is grouped while
the top level of cli/actions stays one file per action. Update import paths.
